backend/requirements.txt
fastapi
uvicorn
anthropic
supabase
groq
python-dotenv
pydantic